fix: :bug: move deleteListUsingID function to List.svelte because I need to use a Svelte event dispatcher

+++ src/lib/List.svelte
@@ -1,26 +1,33 @@
 <script>
 	import { createEventDispatcher } from "svelte";
+	const dispatch = createEventDispatcher();
 
 	export let list;
 	export let editMode = false;
 
 	// console.log(list);
 	const { listname, itemList, color, _id } = list;
-	const dispatch = createEventDispatcher();
 
-	async function deleteList(id) {
-		const res = await fetch(`/api/lists/${id}`, {
+	// funktion som tar bort lista med ett speciellt ID
+	// ta inte bort fel 😄
+	export async function deleteListUsingID(id) {
+		if (id === null) {
+			console.log(id, "wrong");
+			return;
+		}
+		const res = await fetch(`${API_BASE}/lists/${id}`, {
 			method: "DELETE",
 		});
-		const { list } = await res.json();
-		console.log("result:", list);
 		dispatch("deletedList");
-		return list;
 	}
 </script>
 
 <a
 	class:bg-red-400;!hover-bg-red-500={color === "red"}
+  class:bg-orange-400;!hover-bg-orange-500={color === "orange"}
+  class:bg-yellow-400;!hover-bg-yellow-500={color === "yellow"}
+  class:bg-green-400;!hover-bg-green-500={color === "green"}
+  class:bg-blue-400;!hover-bg-blue-500={color === "blue"}
 	class:bg-purple-400;!hover-bg-purple-500={color === "purple"}
 	class="m-4 flex h-64 w-64 cursor-pointer flex-col items-center rounded bg-slate-300 shadow shadow-black/50 hover:bg-slate-200"
 	href="/details/{_id}">
@@ -28,7 +35,7 @@
 	{#if editMode}
 		<div class="absolute h-64 w-64">
 			<button
-				on:click|preventDefault={async () => await deleteList(_id)}
+				on:click|preventDefault={async () => await deleteListUsingID(_id)}
 				class="i-mdi-delete?bg absolute right--3 top--2 rounded-full bg-slate-400 text-3xl" />
 		</div>
 	{/if}
